// dashboard/src/utils/hasura-api/generated/schemas/suggestedArrayRelationship.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { SuggestedArrayRelationshipTo } from './suggestedArrayRelationshipTo';
import type { SuggestedArrayRelationshipType } from './suggestedArrayRelationshipType';
import type { SuggestedRelationshipElement } from './suggestedRelationshipElement';

export interface SuggestedArrayRelationship {
  type?: SuggestedArrayRelationshipType;
  from?: SuggestedRelationshipElement;
  to?: SuggestedArrayRelationshipTo;
}